Move rollup serialization into rollup dir
