+++ packages/core/src/app/serialization/index.ts
@@ -4,6 +4,4 @@ export * from './ownership-transaction'
 export * from './serializers'
 export * from './state-object'
 export * from './state-update'
-export * from './transfer-new-account'
-export * from './transfer-stored-account'
-export * from './swap-tx'
+export * from './rollup'
