+++ packages/core/src/app/serialization/rollup/transfer-stored-account.ts
@@ -1,6 +1,6 @@
 /* Internal Imports */
-import { abi } from '../../app'
-import { AbiEncodable } from '../../types'
+import { abi } from '../../../app'
+import { AbiEncodable } from '../../../types'
 
 /**
  * Creates a TransferStoredAccount from an encoded TransferStoredAccount.
